Split view methods out of the Search class.
